+++ src/commands/esnfarger.py
@@ -1,0 +1,10 @@
+
+def command_esn_colors(channel, user):
+    respond_to(channel, user,
+               "• ESN Cyan #00aeef\n"
+               + "• ESN Magenta #ec008c\n"
+               + "• ESN Green #7ac143\n"
+               + "• ESN Orange #f47b20\n"
+               + "• ESN Dark Blue #2e3192\n"
+               + "• Black #000000\n"
+               + "• White #ffffff")
