+++ src/commands/esnfont.py
@@ -1,0 +1,3 @@
+
+def command_esn_font(channel, user):
+    respond_to(channel, user, "Display font: Kelson Sans\n" + "Content font: Lato")
