+++ src/commands/standliste.py
@@ -1,0 +1,3 @@
+
+def command_stand_list(channel, user):
+    respond_to(channel, user, os.environ.get("STAND_LIST"))
